+++ addons/yat/src/enums/EStringConversionResult.cs
@@ -1,0 +1,8 @@
+namespace YAT.Enums;
+
+public enum EStringConversionResult
+{
+	Success,
+	Invalid,
+	OutOfRange
+}