+++ addons/yat/src/types/CommandData.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading;
+using Godot;
 using YAT.Interfaces;
 using YAT.Scenes;
 
@@ -10,8 +11,8 @@ namespace YAT.Types
 		BaseTerminal Terminal,
 		ICommand Command,
 		string[] RawData,
-		Dictionary<string, object> Arguments,
-		Dictionary<string, object> Options,
+		Dictionary<StringName, object> Arguments,
+		Dictionary<StringName, object> Options,
 		CancellationToken CancellationToken
 	);
 }