+++ addons/yat/src/types/CommandInputType.cs
@@ -1,0 +1,5 @@
+using Godot;
+
+namespace YAT.Types;
+
+public record CommandInputType(StringName Type, float Min, float Max, bool IsArray);