+++ script_templates/Node/ExtensibleCommand.cs
@@ -7,13 +7,19 @@ using YAT.Classes;
 using YAT.Interfaces;
 using YAT.Types;
 
-[Command("_CLASS_", "Lorem ipsum dolor sit amet.", "[b]Usage[/b]: _CLASS_ [i]action[/i]")]
+namespace YAT.Commands;
+
+[Command("_CLASS_")]
+[Usage("_CLASS_ [i]action[/i]")]
+[Description("Lorem ipsum dolor sit amet.")]
 [Argument("action", "string", "The name of the action to run.")]
 public partial class _CLASS_ : Extensible, ICommand
 {
 	public CommandResult Execute(CommandData data)
 	{
 		var extensions = GetCommandExtensions("_CLASS_");
+
+		if (extensions is null) return ICommand.Failure("No extensions found.");
 
 		if (extensions.TryGetValue((string)data.Arguments["action"], out Type extension))
 			return ExecuteExtension(extension, data with { RawData = data.RawData[1..] });