+++ script_templates/Node/Extension.cs
@@ -5,8 +5,10 @@ using YAT.Attributes;
 using YAT.Interfaces;
 using YAT.Types;
 
-[Extension("_CLASS_", "Lorem ipsum dolor sit amet.", "[b]Usage[/b]: _CLASS_")]
-public partial class _CLASS_ : IExtension
+[Extension("_CLASS_")]
+[Usage("_CLASS_")]
+[Description("Lorem ipsum dolor sit amet.")
+public sealed class _CLASS_ : IExtension
 {
 	public CommandResult Execute(CommandData data)
 	{