+++ CHANGELOG.md
@@ -10,16 +10,25 @@ All notable changes to this project will be documented in this file.
 -   More results to ECommandResult.
 -   CommandResult record.
 -   Static methods to ICommand that corresponds to the ECommandResult results.
--   ESceneChangeFailureReason enum.
 -   Networking helper class.
 -   NetworkingOptions record.
 -   -limit option to ping command.
+-   CommandInputType record.
+-   TryParseCommandInputType method for Text helper class.
+-   String type in arguments and options can have defined min and max length.
+-   Enums:
+    -   ESceneChangeFailureReason
+    -   EStringConversionResult
 -   Commands:
     -   TraceRoute
     -   Load
 -   Attributes:
     -   Description
     -   Usage
+    -   CommandInput
+-   Messages:
+    -   InvalidOption,
+    -   ValueOutOfRange
 
 ### Changed
 
@@ -27,10 +36,23 @@ All notable changes to this project will be documented in this file.
 -   Commands now return CommandResult record.
 -   The SceneAboutToChange signal in the Cs command now also gives the old path, not just the new one.
 -   Renamed ping command options.
+-   Argument & Option attributes inherit from CommandInput attribute.
+-   Options with null type are no longer supported.
+-   The way of defining the type for arguments and options has changed.
+-   CommandData now uses StringName instead of string for dictionaries.
+-   Methods StartsWith & EndsWith from Text class are now extensions of string type.
+-   Using default value for bool type options is no longer necessary.
+-   Double is no longer a valid type for options & arguments.
+-   Updated AUTOMATIC_INPUT_VALIDATION.md & script templates.
 
 ### Removed
 
 -   FailureReason enum.
+
+### Fixed
+
+-   It is possible to call the history command via itself, which can cause an endless loop.
+-   The set command throws an exception when there are no registered extensions.
 
 ## [1.23.0-beta 2024-02-13]